# openapi.yaml
openapi: 3.0.3
info:
  title: Cognipeer Console Client API
  version: 1.0.0
  description: |
    Multi-tenant SaaS platform for AI and Agentic services providing LLM services, 
    agent orchestration, vector stores, workflow automation, and analytics.
    
    ## Authentication
    All endpoints require Bearer token authentication using API tokens.
    Include the token in the Authorization header: `Authorization: Bearer <token>`
    
    ## Base URL
    Managed service: `https://api.cognipeer.com/api/client/v1`
    Self-hosted: `http://localhost:3000/api/client/v1`

    ## License
    This API surface is documented from the AGPL community edition of Cognipeer Console.
    Commercial licensing and managed service terms may differ from self-hosted community use.
    
  contact:
    name: Cognipeer Support
    url: https://github.com/Cognipeer/cognipeer-console
  license:
    name: AGPL-3.0-only
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:3000/api/client/v1
    description: Development server
security:
  - BearerAuth: []
...
paths:
...
  /vector/providers/{providerKey}/indexes/{externalId}:
    get:
      tags:
        - Vector Indexes
      summary: Get vector index
      description: Retrieve details of a specific vector index
      operationId: getVectorIndex
      parameters:
        - $ref: '#/components/parameters/ProviderKey'
        - $ref: '#/components/parameters/ExternalId'
      responses:
        '200':
          description: Index details
          content:
            application/json:
              schema:
                type: object
                properties:
                  index:
                    $ref: '#/components/schemas/VectorIndex'
                  provider:
                    $ref: '#/components/schemas/VectorProvider'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    patch:
      tags:
        - Vector Indexes
      summary: Update vector index
      description: Update index name and/or metadata
      operationId: updateVectorIndex
      parameters:
        - $ref: '#/components/parameters/ProviderKey'
        - $ref: '#/components/parameters/ExternalId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVectorIndexRequest'
            example:
              name: "products-v2"
              metadata:
                version: "2.0"
      responses:
        '200':
          description: Index updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  index:
                    $ref: '#/components/schemas/VectorIndex'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags:
        - Vector Indexes
      summary: Delete vector index
      description: Delete a vector index and all its data
      operationId: deleteVectorIndex
      parameters:
        - $ref: '#/components/parameters/ProviderKey'
        - $ref: '#/components/parameters/ExternalId'
      responses:
        '200':
          description: Index deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: API token for authentication
  parameters:
    ProviderKey:
      name: providerKey
      in: path
      required: true
      schema:
        type: string
      description: Unique provider identifier
    ExternalId:
      name: externalId
      in: path
      required: true
      schema:
        type: string
      description: External index identifier
  schemas:
...
    VectorProvider:
      type: object
      properties:
        _id:
          type: string
        key:
          type: string
        driver:
          type: string
          example: "pinecone"
        label:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, inactive, error]
        credentials:
          type: object
          description: Provider-specific credentials (encrypted)
        settings:
          type: object
        metadata:
          type: object
        capabilities:
          type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    VectorIndex:
      type: object
      properties:
        _id:
          type: string
        key:
          type: string
        indexId:
          type: string
        name:
          type: string
        dimension:
          type: integer
        metric:
          type: string
          enum: [cosine, euclidean, dotproduct]
        providerKey:
          type: string
        metadata:
          type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UpdateVectorIndexRequest:
      type: object
      properties:
        name:
          type: string
        metadata:
          type: object
...
    Error:
      type: object
      properties:
        error:
          oneOf:
            - type: string
            - type: object
              properties:
                message:
                  type: string
                type:
                  type: string
                  enum: [invalid_request_error, server_error]
  responses:
    BadRequest:
      description: Bad request - Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - Invalid or missing API token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Invalid API token"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Not found"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              message: "Internal server error"
              type: "server_error"